refactor(test): extract MemSandbox to conftest for shared use across sync e2e
