+++ backend/tests/e2e/conftest.py
@@ -42,14 +42,94 @@ from cubebox.repositories import (
     OrganizationRepository,
     WorkspaceRepository,
 )
+from cubebox.sandbox.base import ExecuteResult, Sandbox
 from cubebox.sandbox.lazy import LazySandbox
 from cubebox.sandbox.manager import SandboxManager
+from cubebox.skills.sandbox_paths import SKILLS_ROOT
 from tests.e2e.helpers import csrf_cookie_name
 
 
 def _auth_cookie_name() -> str:
     """Resolved auth cookie name; honours per-worktree env override."""
     return str(_cubebox_config.get("auth.cookie_name", "cubebox_auth"))
+
+
+# ---------------------------------------------------------------------------
+# MemSandbox: shared in-memory sandbox for skills sync tests
+# ---------------------------------------------------------------------------
+
+
+class MemSandbox(Sandbox):
+    """Minimal in-memory Sandbox for testing ``_sync_skills``.
+
+    ``upload`` stores bytes by path; ``download`` reads them back.
+    ``execute`` handles the ``tar -xzf /tmp/skills_delta.tgz -C <root>``
+    command that ``_sync_skills`` emits — it unpacks the tarball into the
+    in-memory FS so SKILL.md entries are readable afterwards.  All other
+    commands are accepted silently (mkdir -p, rm -f etc. are benign no-ops
+    in the test context).
+    """
+
+    _SKILLS_TGZ = "/tmp/skills_delta.tgz"
+
+    def __init__(self) -> None:
+        self._files: dict[str, bytes] = {}
+
+    @property
+    def id(self) -> str:
+        return "mem-sandbox"
+
+    @property
+    def workdir(self) -> str:
+        return "/workspace"
+
+    async def execute(
+        self,
+        command: str,
+        *,
+        timeout: int | None = None,
+        envs: dict[str, str] | None = None,
+    ) -> ExecuteResult:
+        del timeout, envs
+        # Handle the tar extract command that _sync_skills emits.
+        # Format: "mkdir -p ... && rm -rf ... && tar -xzf /tmp/skills_delta.tgz -C /workspace/.skills && ..."
+        if "tar -xzf" in command and self._SKILLS_TGZ in command:
+            tgz = self._files.get(self._SKILLS_TGZ)
+            if tgz:
+                # Extract into ``{SKILLS_ROOT}/<path>`` entries.
+                import tarfile
+
+                with tarfile.open(fileobj=io.BytesIO(tgz), mode="r:gz") as tf:
+                    for member in tf.getmembers():
+                        if member.isfile():
+                            f = tf.extractfile(member)
+                            if f is not None:
+                                dest = f"{SKILLS_ROOT}/{member.name}"
+                                self._files[dest] = f.read()
+                # Remove the tgz (mirrors the real command's ``rm -f``).
+                self._files.pop(self._SKILLS_TGZ, None)
+        # rm -rf / rm -f: drop matching keys.
+        for token in command.split("&&"):
+            stripped = token.strip()
+            if stripped.startswith("rm -rf ") or stripped.startswith("rm -f "):
+                path = stripped.split(None, 2)[-1].strip("'\"")
+                self._files.pop(path, None)
+        return ExecuteResult(output="", exit_code=0)
+
+    async def upload(self, files: list[tuple[str, bytes]]) -> None:
+        for path, content in files:
+            self._files[path] = content
+
+    async def download(self, paths: list[str]) -> list[tuple[str, bytes]]:
+        result: list[tuple[str, bytes]] = []
+        for path in paths:
+            if path not in self._files:
+                raise FileNotFoundError(path)
+            result.append((path, self._files[path]))
+        return result
+
+    async def close(self) -> None:
+        pass
 
 
 def pytest_collection_modifyitems(config: pytest.Config, items: list[pytest.Item]) -> None:
